pr feedback from Ryan
