+++ Sources/ArcGISToolkit/Components/Authentication/Authenticator.swift
@@ -88,64 +88,3 @@ extension Authenticator: NetworkAuthenticationChallengeHandler {
         return await challengeContinuation.value
     }
 }
-
-public extension AuthenticationManager {
-    /// Sets up authenticator as ArcGIS and Network challenge handlers to handle authentication
-    /// challenges.
-    /// - Parameter authenticator: The authenticator to be used for handling challenges.
-    func handleAuthenticationChallenges(using authenticator: Authenticator) {
-        arcGISAuthenticationChallengeHandler = authenticator
-        networkAuthenticationChallengeHandler = authenticator
-    }
-    
-    /// Sets up new credential stores that will be persisted to the keychain.
-    /// - Remark: The credentials will be stored in the default access group of the keychain.
-    /// You can find more information about what the default group would be here:
-    /// https://developer.apple.com/documentation/security/keychain_services/keychain_items/sharing_access_to_keychain_items_among_a_collection_of_apps
-    /// - Parameters:
-    ///   - access: When the credentials stored in the keychain can be accessed.
-    ///   - synchronizesWithiCloud: A Boolean value indicating whether the credentials are synchronized with iCloud.
-    func setupPersistentCredentialStorage(
-        access: ArcGIS.KeychainAccess,
-        synchronizesWithiCloud: Bool = false
-    ) async throws {
-        let previousArcGISCredentialStore = arcGISCredentialStore
-        
-        // Set a persistent ArcGIS credential store on the ArcGIS environment.
-        arcGISCredentialStore = try await .makePersistent(
-            access: access,
-            synchronizesWithiCloud: synchronizesWithiCloud
-        )
-        
-        do {
-            // Set a persistent network credential store on the ArcGIS environment.
-            await setNetworkCredentialStore(
-                try await .makePersistent(access: access, synchronizesWithiCloud: synchronizesWithiCloud)
-            )
-        } catch {
-            // If making the shared network credential store persistent fails,
-            // then restore the ArcGIS credential store.
-            arcGISCredentialStore = previousArcGISCredentialStore
-            throw error
-        }
-    }
-    
-    /// Clears all ArcGIS and network credentials from the respective stores.
-    /// Note: This sets up new `URLSessions` so that removed network credentials are respected
-    /// right away.
-    func clearCredentialStores() async {
-        // Revoke tokens for OAuth user credentials.
-        let oAuthUserCredentials = arcGISCredentialStore.credentials.compactMap { $0 as? OAuthUserCredential }
-        oAuthUserCredentials.forEach { oAuthUserCredential in
-            Task {
-                try? await oAuthUserCredential.revokeToken()
-            }
-        }
-        
-        // Clear ArcGIS Credentials.
-        arcGISCredentialStore.removeAll()
-        
-        // Clear network credentials.
-        await networkCredentialStore.removeAll()
-    }
-}
